rename\remove some files
