+++ client/src/pages/src/visualizer-component/types/visualizerType.ts
