--- debugger/src/gdb_scripts/ptype_preprocessed
@@ -1,10 +1,0 @@
-# 0 "/app/src/gdb_scripts/ptype_output.c"
-# 0 "<built-in>"
-# 0 "<command-line>"
-# 1 "/usr/include/stdc-predef.h" 1 3 4
-# 0 "<command-line>" 2
-# 1 "/app/src/gdb_scripts/ptype_output.c"
-struct node {
-    int data;
-    struct node *next;
-};
